requirements.txt
amqp==5.1.1
APScheduler==3.6.3
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
cachetools==4.2.2
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.1
django-admin-sortable==2.3
django-ckeditor==6.5.1
django-colorfield==0.7.2
django-cors-headers==3.13.0
django-js-asset==2.0.0
django-tinymce==3.5.0
djangorestframework==3.14.0
drf-yasg==1.21.4
flower==1.2.0
gevent==22.10.2
greenlet==2.0.1
humanize==4.4.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.2.0
prometheus-client==0.15.0
prompt-toolkit==3.0.33
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
python-telegram-bot==13.14
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
redis==4.3.5
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
tornado==6.1
tzdata==2022.6
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
zope.event==4.5.0
zope.interface==5.5.2
